% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_partition.r
\name{meta_partition}
\alias{meta_partition}
\title{Meta Partition}
\usage{
meta_partition(
  model,
  moderators,
  n.to.small = 8,
  c = 1,
  control = rpart.control(xval = 10, minbucket = 3, minsplit = 6, cp = 1e-04)
)
}
\arguments{
\item{model}{sdfs}

\item{moderators}{sfssdf}

\item{c}{sdfdf}

\item{control}{sdf}

\item{...}{}
}
\value{

}
\description{
Does Meta-Partition as discribed in ..
}
\examples{
data(Method.Study.Data)
Model <-
  meta::metacor(
    ESr, n,
    data = Method.Study.Data,
    studlab = Study,
    comb.random = FALSE
  )

}
\references{
Ortega Z, Martín-Vallejo J, Mencía A,
Galindo-Villardón MP, Pérez-Mellado V (2016)
Introducing Meta-Partition, a Useful Methodology to
Explore Factors That Influence Ecological Effect
Sizes. PLoS ONE 11(7): e0158624. doi:10.1371/
  journal.pone.0158624
}
